moving StarTableAdapter classes to IrisCommon
