+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/plotter/PlotterView.java
@@ -39,8 +39,8 @@ import cfa.vo.iris.IrisApplication;
 import cfa.vo.iris.gui.GUIUtils;
 import cfa.vo.iris.gui.NarrowOptionPane;
 import cfa.vo.iris.sed.ExtSed;
-import cfa.vo.iris.visualizer.stil.SegmentStarTableAdapter;
-import cfa.vo.iris.visualizer.stil.StarTableAdapter;
+import cfa.vo.iris.sed.stil.SegmentStarTableAdapter;
+import cfa.vo.iris.sed.stil.StarTableAdapter;
 import cfa.vo.iris.visualizer.stil.StilPlotter;
 import cfa.vo.sedlib.ISegment;
 import cfa.vo.iris.IWorkspace;
